Refactor KEYB out of dos_programs
