+++ src/dos/program_keyb.cpp
@@ -1,0 +1,87 @@
+/*
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ *  Copyright (C) 2020-2021  The DOSBox Staging Team
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+
+#include "dos_keyboard_layout.h"
+#include "program_keyb.h"
+#include "string_utils.h"
+
+void KEYB::Run(void) {
+	if (cmd->FindCommand(1,temp_line)) {
+		if (cmd->FindString("?",temp_line,false)) {
+			WriteOut(MSG_Get("PROGRAM_KEYB_SHOWHELP"));
+		} else {
+			/* first parameter is layout ID */
+			Bitu keyb_error=0;
+			std::string cp_string;
+			Bit32s tried_cp = -1;
+			if (cmd->FindCommand(2,cp_string)) {
+				/* second parameter is codepage number */
+				tried_cp=atoi(cp_string.c_str());
+				char cp_file_name[256];
+				if (cmd->FindCommand(3,cp_string)) {
+					/* third parameter is codepage file */
+					safe_strcpy(cp_file_name, cp_string.c_str());
+				} else {
+					/* no codepage file specified, use automatic selection */
+					safe_strcpy(cp_file_name, "auto");
+				}
+
+				keyb_error=DOS_LoadKeyboardLayout(temp_line.c_str(), tried_cp, cp_file_name);
+			} else {
+				keyb_error=DOS_SwitchKeyboardLayout(temp_line.c_str(), tried_cp);
+			}
+			switch (keyb_error) {
+				case KEYB_NOERROR:
+					WriteOut(MSG_Get("PROGRAM_KEYB_NOERROR"),temp_line.c_str(),dos.loaded_codepage);
+					break;
+				case KEYB_FILENOTFOUND:
+					WriteOut(MSG_Get("PROGRAM_KEYB_FILENOTFOUND"),temp_line.c_str());
+					WriteOut(MSG_Get("PROGRAM_KEYB_SHOWHELP"));
+					break;
+				case KEYB_INVALIDFILE:
+					WriteOut(MSG_Get("PROGRAM_KEYB_INVALIDFILE"),temp_line.c_str());
+					break;
+				case KEYB_LAYOUTNOTFOUND:
+					WriteOut(MSG_Get("PROGRAM_KEYB_LAYOUTNOTFOUND"),temp_line.c_str(),tried_cp);
+					break;
+				case KEYB_INVALIDCPFILE:
+					WriteOut(MSG_Get("PROGRAM_KEYB_INVCPFILE"),temp_line.c_str());
+					WriteOut(MSG_Get("PROGRAM_KEYB_SHOWHELP"));
+					break;
+				default:
+				        LOG(LOG_DOSMISC, LOG_ERROR)("KEYB:Invalid returncode %x",
+				         static_cast<uint32_t>(keyb_error));
+				        break;
+			}
+		}
+	} else {
+		/* no parameter in the command line, just output codepage info and possibly loaded layout ID */
+		const char* layout_name = DOS_GetLoadedLayout();
+		if (layout_name==NULL) {
+			WriteOut(MSG_Get("PROGRAM_KEYB_INFO"),dos.loaded_codepage);
+		} else {
+			WriteOut(MSG_Get("PROGRAM_KEYB_INFO_LAYOUT"),dos.loaded_codepage,layout_name);
+		}
+	}
+}
+
+void KEYB_ProgramStart(Program * * make) {
+	*make=new KEYB;
+}
